+++ README.md
@@ -70,8 +70,8 @@
 
 ## 몬스터           + 페이즈 방식 , MP 없는 걸로, 보스 방어력스탯은 평균수준정도로
 
-#          NAME  |  SKILL(MP-30)                   일정 페이즈: 스탯창에 MP불필요, 하지만 수식 더 필요함..
-#                                                  랜덤 발동 : 보스 스테이터스에 MP포함, 랜덤 확률로 보스 스킬발동
+#          NAME  |  SKILL(MP-30)         일정 페이즈: 스탯창에 MP불필요, 하지만 수식 더 필요함..
+#                                     랜덤발동 : 보스 스테이터스에 MP포함, 랜덤 확률로 보스스킬발동
 #                       5턴
 - 중간보스1: 슈뢰딩거 | 슈뢰딩거의상 자식
                 랜덤 1마리 1~3턴간 CC / 랜덤타깃딜링
